src/requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
azure-appconfiguration-provider==2.4.0
azure-identity==1.19.0
aiohttp
